+++ build.sbt
@@ -25,7 +25,7 @@ scalacOptions := Seq(
   //"-opt-inline-from:<sources>",
 ) else Nil)
 
-sbtPlugin := true
+enablePlugins(SbtPlugin)
 
 scriptedBufferLog := false
 
@@ -41,12 +41,15 @@ scriptedLaunchOpts ++= Seq("-Xmx1024M", "-Dplugin.version=" + version.value)
 crossSbtVersions := Vector("0.13.18", "1.1.6")
 
 val amazonSDKVersion = "1.12.129"
+val testcontainersScalaVersion = "0.27.0"
 
 libraryDependencies ++= Seq(
   "com.amazonaws" % "aws-java-sdk-s3" % amazonSDKVersion,
   "com.amazonaws" % "aws-java-sdk-sts" % amazonSDKVersion,
-  "org.apache.ivy" % "ivy" % "2.4.0",
-  "org.scalatest" %% "scalatest" % "3.2.10" % Test
+  "org.apache.ivy" % "ivy" % "2.5.0",
+  "org.scalatest" %% "scalatest" % "3.2.10" % Test,
+  "com.dimafeng" %% "testcontainers-scala-scalatest" % testcontainersScalaVersion % Test,
+  "com.dimafeng" %% "testcontainers-scala-localstack-v2" % testcontainersScalaVersion % Test
 )
 
 // Tell the sbt-release plugin to use publishSigned
@@ -82,7 +85,7 @@ releaseProcess := Seq[ReleaseStep](
   pushChanges
 )
 
-publishArtifact in Test := false
+Test / publishArtifact := false
 
 pomIncludeRepository := { _ => false }
