+++ src/sbt-test/fm-sbt-s3-resolver/resolver-style/project/plugins.sbt
@@ -1,0 +1,6 @@
+{
+  val pluginVersion = Option(System.getProperty("plugin.version")) getOrElse {
+    throw new RuntimeException("The system property 'plugin.version' is not defined. Specify this property using the scriptedLaunchOpts -D.")
+  }
+  addSbtPlugin("com.frugalmechanic" % "fm-sbt-s3-resolver" % pluginVersion)
+}
