+++ src/sbt-test/fm-sbt-s3-resolver/resolver-style/test
@@ -1,0 +1,4 @@
+> mavenStyle/check
+# GitHub doesn't support ivyStyle
+#> ivyStyle/check
+> clean
